Improve UT Coverage for TF 3x
