+++ neural_compressor/tensorflow/algorithms/static_quant/tensorflow.py
@@ -37,7 +37,6 @@ from neural_compressor.tensorflow.utils import (
     BaseDataLoader,
     BaseModel,
     CpuInfo,
-    Dequantize,
     Statistics,
     deep_get,
     dump_elapsed_time,
@@ -87,8 +86,8 @@ class TensorFlowAdaptor:
         cfg_yaml_name = "{}.yaml".format(self.__class__.__name__[: -len("Adaptor")].lower())
         self.itex_mode = self.backend == "itex" or cfg_yaml_name == "tensorflow_itex.yaml"
 
-        # if self.itex_mode:
-        #     self._check_itex()
+        if self.itex_mode:
+            self._check_itex()
 
         self.query_handler = TensorflowQuery(
             local_config_file=os.path.join(os.path.dirname(__file__), cfg_yaml_name),
@@ -118,284 +117,6 @@ class TensorFlowAdaptor:
                 "The Intel® Extension for TensorFlow is not installed. "
                 "Please install it to run models on ITEX backend"
             )
-
-    def _log_histogram(self, writer, tag, values, step=0, bins=1000):
-        """Writes a histogram for later analysis."""
-        # Convert to a numpy array
-        values = np.array(values)
-
-        # Create and write Summary
-        # update using TF2.X API
-        with writer.as_default():
-            tf.summary.histogram(tag, values, step)
-            writer.flush()
-
-    def _pre_hook_for_hvd(self, dataloader=None):
-        """Pre hook for Horovod."""
-        import horovod.tensorflow as hvd
-
-        self.hvd = hvd
-        self.hvd.init()
-
-    @dump_elapsed_time(customized_msg="Model inference")
-    def evaluate(
-        self,
-        model,
-        dataloader,
-        postprocess=None,
-        metrics=None,
-        measurer=None,
-        iteration=-1,
-        tensorboard=False,
-        fp32_baseline=False,
-    ):
-        """Evaluate the model for specified metric on validation dataset.
-
-        Args:
-            model ([Graph, GraphDef or Path String]): The model could be the graph,
-                        graph_def object, the frozen pb or ckpt/savedmodel folder path.
-            dataloader (generator): generate the data and labels.
-            postprocess (object, optional): process the result from the model
-            metrics (list, optional): Depends on model category. Defaults to None.
-            measurer (object, optional): for precise benchmark measurement.
-            iteration(int, optional): control steps of mini-batch
-            tensorboard (boolean, optional): for tensorboard inspect tensor.
-            fp32_baseline (boolean, optional): only for compare_label=False pipeline
-
-        Returns:
-            [float]: evaluation result, the larger is better.
-        """
-        from neural_compressor.tensorflow.quantization.utils.utility import iterator_sess_run
-
-        outputs = model.output_tensor_names
-
-        if getattr(dataloader, "distributed", False):
-            import horovod.tensorflow as hvd
-
-            hvd.init()
-            # If metric.hvd is not None then run distributed inference
-            for metric in metrics:
-                metric.hvd = hvd
-            try:
-                len_dataloader = len(dataloader)
-            except:
-                logger.info(
-                    "The length of the distributed evaluation dataloader is unknown."
-                    "When the iteration of evaluation dataloader in each process is "
-                    "inconsistent, an error may occur."
-                )
-            else:
-                list_len_dataloader = hvd.allgather_object(len_dataloader)
-                if hvd.rank() == 0:
-                    for i in range(len(list_len_dataloader) - 1):
-                        if list_len_dataloader[i] != list_len_dataloader[i + 1]:
-                            raise AttributeError(
-                                "The evaluation dataloader's iteration is"
-                                "different between processes, please reset dataloader's batch_size."
-                            )
-            logger.info(
-                "Rank {!s} dataloaders' data distribution balance check for evaluation have been finished.".format(
-                    hvd.allgather_object(hvd.rank())
-                )
-            )
-        if tensorboard:
-            from tensorflow.python.framework import tensor_util
-
-            from neural_compressor.tensorflow.quantization.utils.graph_util import GraphAnalyzer
-
-            output_postfix = "_fp32.output"
-            inspect_node_types = [
-                "Conv2D",
-                "DepthwiseConv2dNative",
-                "MaxPool",
-                "AvgPool",
-                "ConcatV2",
-                "MatMul",
-                "FusedBatchNormV3",
-                "FusedBatchNorm",
-                "BiasAdd",
-                "_MklFusedInstanceNorm",
-                "Relu",
-                "Relu6",
-                "Dequantize",
-            ]
-            fp32_inspect_node_name = []
-            int8_inspect_node_name = []
-            q_node_scale = {}
-            if self.dump_times == 0:
-                temp_dir = "./runs/eval/baseline"
-            else:
-                temp_dir = "./runs/eval/tune_" + str(self.dump_times)
-            if os.path.isdir(temp_dir):
-                import shutil
-
-                shutil.rmtree(temp_dir, ignore_errors=True)
-            # Create the writer using TF2.x APIs to handle eager executions
-            writer = tf.summary.create_file_writer(temp_dir)  # pylint: disable=no-member
-            with writer.as_default():
-                tf.summary.graph(model.graph)  # pylint: disable=no-member
-
-            cur_graph = GraphAnalyzer()
-            cur_graph.graph = model.graph_def
-            cur_graph.parse_graph()
-            graph_info = cur_graph.node_name_details
-            for node in model.graph_def.node:
-                if node.op in inspect_node_types:
-                    fp32_inspect_node_name.append(node.name)
-                # Tensor dump supported quantized op including,
-                # Requantize, QuantizedConv2DAndRequantize,
-                # QuantizedConv2DAndReluAndRequantize,
-                # QuantizedConv2DWithBiasAndRequantize,
-                # QuantizedConv2DWithBiasAndReluAndRequantize,
-                # QuantizedConv2DWithBiasSignedSumAndReluAndRequantize,
-                # QuantizedConv2DWithBiasSumAndReluAndRequantize,
-                # QuantizedDepthwiseConv2DWithBiasAndReluAndRequantize,
-                # QuantizedMatMulWithBiasAndReluAndRequantize,
-                # QuantizedMatMulWithBiasAndRequantize
-                elif node.op.find("Requantize") != -1:
-                    out_min = -2
-                    out_max = -1
-                    if node.op.find("Sum") != -1:
-                        out_min = -5
-                        out_max = -4
-                    q_out_min = graph_info[node.input[out_min]].node.attr["value"].tensor.float_val[0]
-                    q_out_max = graph_info[node.input[out_max]].node.attr["value"].tensor.float_val[0]
-                    q_node_scale[node.name] = (node.op, q_out_min, q_out_max)
-                    int8_inspect_node_name.append(node.name)
-                # Inspect weights, bias. Need further optimize
-                if node.op == "Const" and graph_info[graph_info[node.name].outputs[0]].node.op in [
-                    "Conv2D",
-                    "DepthwiseConv2dNative",
-                    "MatMul",
-                    "FusedBatchNormV3",
-                    "_MklFusedInstanceNorm",
-                    "BiasAdd",
-                ]:
-                    const_value = tensor_util.MakeNdarray(node.attr.get("value").tensor).astype(np.float32)
-                    self._log_histogram(writer, node.name, const_value)
-
-            outputs.extend(fp32_inspect_node_name)
-            if len(int8_inspect_node_name) > 0:
-                output_postfix = "_int8.output"
-                outputs.extend(int8_inspect_node_name)
-
-        if metrics:
-            for metric in metrics:
-                metric.reset()
-            self.fp32_preds_as_label = any(
-                [hasattr(metric, "compare_label") and not metric.compare_label for metric in metrics]
-            )
-
-        origin_output_tensor_names = model.output_tensor_names
-        model.output_tensor_names = outputs
-        input_tensor = model.input_tensor
-        output_tensor = model.output_tensor if len(model.output_tensor) > 1 else model.output_tensor[0]
-        logger.info("Start to evaluate the TensorFlow model.")
-
-        def eval_func(dataloader):
-            results = []
-            for idx, (inputs, labels) in enumerate(dataloader):
-                # dataloader should keep the order and len of inputs same with input_tensor
-                if len(input_tensor) == 1:
-                    feed_dict = {}
-                    if isinstance(inputs, dict) or isinstance(inputs, OrderedDict) or isinstance(inputs, UserDict):
-                        for name in inputs:
-                            for tensor in input_tensor:
-                                pos = tensor.name.rfind(":")
-                                t_name = tensor.name if pos < 0 else tensor.name[:pos]
-                                if name == t_name:
-                                    feed_dict[tensor] = inputs[name]
-                                    break
-                    else:
-                        feed_dict = {input_tensor[0]: inputs}  # get raw tensor using index [0]
-                else:
-                    assert len(input_tensor) == len(inputs), "inputs len must equal with input_tensor"
-                    feed_dict = {}
-                    if isinstance(inputs, dict) or isinstance(inputs, OrderedDict) or isinstance(inputs, UserDict):
-                        for name in inputs:
-                            for tensor in input_tensor:
-                                pos = tensor.name.rfind(":")
-                                t_name = tensor.name if pos < 0 else tensor.name[:pos]
-                                if name == t_name:
-                                    feed_dict[tensor] = inputs[name]
-                                    break
-                    else:
-                        feed_dict = dict(zip(input_tensor, inputs))
-
-                if model.iter_op:
-                    predictions = iterator_sess_run(
-                        model.sess, model.iter_op, feed_dict, output_tensor, iteration, measurer
-                    )
-                elif measurer is not None:
-                    measurer.start()
-                    predictions = model.sess.run(output_tensor, feed_dict)
-                    measurer.end()
-                else:
-                    predictions = model.sess.run(output_tensor, feed_dict)
-
-                if self.fp32_preds_as_label:
-                    self.fp32_results.append(predictions) if fp32_baseline else results.append(predictions)
-
-                # Inspect node output, just get 1st iteration output tensors for now
-                if idx == 0 and tensorboard:
-                    for index, node_name in enumerate(outputs):
-                        tensor = predictions[index]
-                        if node_name in int8_inspect_node_name:
-                            tensor = Dequantize(predictions[index], q_node_scale[node_name])
-                        self._log_histogram(writer, node_name + output_postfix, tensor.astype(np.float32), idx)
-                    writer.close()
-                if isinstance(predictions, list):
-                    if len(origin_output_tensor_names) == 1:
-                        predictions = predictions[0]
-                    elif len(origin_output_tensor_names) > 1:
-                        predictions = predictions[: len(origin_output_tensor_names)]
-                if postprocess is not None:
-                    predictions, labels = postprocess((predictions, labels))
-                if metrics:
-                    for metric in metrics:
-                        if not hasattr(metric, "compare_label") or (
-                            hasattr(metric, "compare_label") and metric.compare_label
-                        ):
-                            metric.update(predictions, labels)
-                if idx + 1 == iteration:
-                    break
-            return results
-
-        if isinstance(dataloader, BaseDataLoader) and not self.benchmark:
-            try:
-                results = eval_func(dataloader)
-            except Exception:  # pragma: no cover
-                logger.warning("Fail to forward with batch size={}, set to {} now.".format(dataloader.batch_size, 1))
-                dataloader.batch(1)
-                results = eval_func(dataloader)
-        else:  # pragma: no cover
-            results = eval_func(dataloader)
-
-        if self.fp32_preds_as_label:
-            from neural_compressor.tensorflow.quantization.utils.utility import collate_tf_preds
-
-            if fp32_baseline:
-                results = collate_tf_preds(self.fp32_results)
-                reference = results
-            else:
-                reference = collate_tf_preds(self.fp32_results)
-                results = collate_tf_preds(results)
-            for metric in metrics:
-                if hasattr(metric, "compare_label") and not metric.compare_label:
-                    metric.update(results, reference)
-
-        acc = 0 if metrics is None else [metric.result() for metric in metrics]
-        if tensorboard:
-            new_dir = temp_dir + "_acc_" + str(acc)
-            writer.close()
-            if os.path.isdir(new_dir):
-                import shutil
-
-                shutil.rmtree(new_dir, ignore_errors=True)
-            os.rename(temp_dir, new_dir)
-            self.dump_times += 1
-        model.output_tensor_names = origin_output_tensor_names
-        return acc if not isinstance(acc, list) or len(acc) > 1 else acc[0]
 
     def _tuning_cfg_to_fw(self, tuning_cfg):
         """Parse the neural_compressor wrapped configuration to Tensorflow.
@@ -468,21 +189,11 @@ class TensorFlowAdaptor:
         """
         assert (
             self.approach != "post_training_dynamic_quant"
-        ), "Dynamic quantization is not supported on TensorFlow framework now!"
+        ), "Dynamic Quantization is not supported on TensorFlow framework now!"
 
-        if self.approach == "quant_aware_training":  # pragma: no cover
-            assert (
-                q_func is not None
-            ), "quantization aware training mode \
-                is not configured correctly"
-
-            from neural_compressor.tensorflow.utils import Model
-
-            qat_model = q_func(model)
-
-            return self.convert(Model(qat_model), "QAT", "default")
-
-        assert q_func is None, "post-training quantization mode is not support calibration function for Tensorflow!"
+        assert (
+            self.approach != "quant_aware_training"
+        ), "Quantize Aware Training is not supported on TensorFlow framework now!"
 
         self.calib_sampling_size = calib_dataloader.batch_size * calib_iteration
         tune_cfg = self.parse_quant_config(quant_config, model, calib_iteration)
@@ -612,7 +323,7 @@ class TensorFlowAdaptor:
                 continue
             possible_int8_res = [name for name in int8_op_prefix_list if i.op.find(name) != -1]
 
-            if any(possible_int8_res):
+            if any(possible_int8_res):  # pragma: no cover
                 origin_op_type = possible_int8_res[0].split("Quantized")[-1]
                 if origin_op_type == "FusedBatchNorm":
                     origin_op_type = "FusedBatchNormV3"
@@ -1019,263 +730,6 @@ class TensorFlowAdaptor:
                 g.replace_constant_graph_with_constant_node(min_node, current_node.input[5])
                 g.replace_constant_graph_with_constant_node(max_node, current_node.input[6])
 
-    def inspect_weight_and_bias(self, node_list, graph_def, graph_info, graph_node_name_mapping):
-        """Inspect the weights and biases."""
-        from neural_compressor.tensorflow.quantization.utils.utility import (
-            get_tensor_val_from_graph_node,
-            int8_node_name_reverse,
-        )
-        from neural_compressor.tensorflow.utils import dequantize_weight
-
-        weights_result = {}
-        inspect_nodes = []
-        node_set = set(node_list)
-        for node in graph_def.node:
-            node_name = node.name
-            if "Quantized" in node.op:
-                node_name = int8_node_name_reverse(node)
-            if node_name in node_set and ("Conv" in node.op or "Mul" in node.op):
-                inspect_nodes.append(node)
-        logger.debug(f"Start to inspect weight and bias for: {[node.name for node in inspect_nodes]}.")
-        for node in inspect_nodes:
-            # inspect weights and bias
-            node_name = node.name
-            weight_node_name = node.input[1]
-            weight_node = graph_node_name_mapping[weight_node_name]
-            if weight_node.op != "Const":  # skip the matmul whose two inputs are previous output
-                continue
-            weight_node_val = get_tensor_val_from_graph_node(graph_node_name_mapping, weight_node_name)
-            weight_node_val = weight_node_val.astype("float32")
-            # dequantize the weight for quantized model
-            if "Quantized" in node.op:
-                node_name = int8_node_name_reverse(node)
-                weight_node_name_pre = weight_node_name.split("_qint8_const")[0]
-                min_filter_node = weight_node_name_pre + "_min"
-                max_filter_node = weight_node_name_pre + "_max"
-                if graph_info[min_filter_node].node.attr["value"].tensor.float_val:
-                    min_filter_val = graph_info[min_filter_node].node.attr["value"].tensor.float_val
-                    max_filter_val = graph_info[max_filter_node].node.attr["value"].tensor.float_val
-                else:
-                    min_filter_val = get_tensor_val_from_graph_node(graph_node_name_mapping, min_filter_node)
-                    max_filter_val = get_tensor_val_from_graph_node(graph_node_name_mapping, max_filter_node)
-                weight_node_val = dequantize_weight(weight_node_val, min_filter_val, max_filter_val)
-            weights_result[node_name] = {weight_node_name: weight_node_val}
-        return weights_result
-
-    def fused_node_mapping(self, node_list, pattern_mapping, graph_info, graph_node_name_mapping):
-        """Create the mapping between first node and last node in fused sequence.
-
-        Args:
-            node_list: node name list
-            pattern_mapping:  key: node name, val: node pattern mapping
-            graph_info: key: node name, val: node details
-            graph_node_name_mapping: key: node name, val: node
-        Returns:
-            fused_mapping: key: first node name in fused seq, val: last node in fused seq
-            fused_mapping_reverse: key: last node in fused seq, val: first node name in fused seq
-        """
-        fused_mapping = {}
-        fused_mapping_reverse = {}
-        for node_name in node_list:
-            fused_seq = pattern_mapping[node_name]["sequence"].split(",")
-            # for the node not fused with others
-            if len(fused_seq) == 1:
-                fused_mapping[node_name] = node_name
-                fused_mapping_reverse[node_name] = node_name
-                continue
-            _next_node_name = node_name
-            for _next_node_op_type in fused_seq[1:]:
-                node_details = graph_info[_next_node_name]
-                for node_output_name in node_details.outputs:
-                    if graph_node_name_mapping[node_output_name].op == "Cast":
-                        cast_node = graph_node_name_mapping[node_output_name]
-                        node_output_name = graph_info[cast_node.name].outputs[0]
-                    if graph_node_name_mapping[node_output_name].op in [_next_node_op_type, "Cast"]:
-                        _next_node_name = node_output_name
-            fused_mapping[node_name] = _next_node_name
-            fused_mapping_reverse[_next_node_name] = node_name
-        return fused_mapping, fused_mapping_reverse
-
-    def _inspect_tensor_inference(self, inspect_node_dict, model, dataloader, iteration_list):
-        """Do inference for inspect activation."""
-        out_tensor_lst = []
-        out_tensor_lst += [{n: [n + ":" + str(i) for i in range(3)]} for n in inspect_node_dict["qreq_node"]]
-        out_tensor_lst += [{n: n + ":0"} for n in inspect_node_dict["qdq_node"]]
-        out_tensor_lst += [{n: n + ":0"} for n in inspect_node_dict["f_node"]]
-        out_cnt = len(out_tensor_lst)
-        iteration_list = set(iteration_list)
-        input_tensor = model.input_tensor
-        logger.info("Start to do inference for inspect activation.")
-        activation_result = []
-        for idx, (inputs, labels) in enumerate(dataloader):
-            model_out = []
-            if idx + 1 > max(iteration_list):
-                break
-            if idx + 1 not in iteration_list:
-                continue
-            if len(input_tensor) == 1:
-                feed_dict = {input_tensor[0]: inputs}  # get raw tensor using index [0]
-            else:
-                assert len(input_tensor) == len(inputs), "inputs len must equal with input_tensor"
-                feed_dict = dict(zip(input_tensor, inputs))
-            # TODO find an optimized method to avoid multiple runs
-            for i, out_t in enumerate(out_tensor_lst):
-                logger.debug(f"Finished inspect {i}/{out_cnt} nodes, current inspect node {out_t.keys()}.")
-                model_out.append(model.sess.run(out_t, feed_dict))
-            activation_result.append(model_out)
-        return activation_result
-
-    def inspect_activation(
-        self, node_list, graph_def, graph_node_name_mapping, quantization_cfg, dataloader, iteration_list, graph_info
-    ):
-        """Inspect the activation."""
-        from neural_compressor.tensorflow.utils import Model
-
-        original_graph_node_mapping = {}
-        for node in graph_def.node:
-            original_graph_node_mapping[node.name] = node
-        inspect_node_dict = {"qdq_node": [], "qreq_node": [], "f_node": []}
-        for node_name in node_list:
-            node = graph_node_name_mapping[node_name]
-            if "Quantized" in node.op and "Dequantize" in node.op:
-                inspect_node_dict["qdq_node"].append(node.name)
-            elif "Quantized" in node.op or "_Quantized" in node.op or "Requantize" in node.op:
-                inspect_node_dict["qreq_node"].append(node.name)
-            else:
-                inspect_node_dict["f_node"].append(node_name)
-        pattern_mapping = {}
-        node_dict = quantization_cfg["op"]
-        for node_name_and_type in node_dict.keys():
-            node_name, _ = node_name_and_type
-            if "pattern" in node_dict[node_name_and_type]:
-                pattern_mapping[node_name] = node_dict[node_name_and_type]["pattern"]
-            else:
-                pattern_mapping[node_name] = {"sequence": node_name}
-        if inspect_node_dict["f_node"]:
-            fuse_map, fuse_map_reverse = self.fused_node_mapping(
-                inspect_node_dict["f_node"], pattern_mapping, graph_info, graph_node_name_mapping
-            )
-            inspect_node_dict["f_node"] = [fuse_map[n] for n in inspect_node_dict["f_node"]]
-        # build model and do inference
-        model = Model(graph_def)
-        activation_result = self._inspect_tensor_inference(inspect_node_dict, model, dataloader, iteration_list)
-        final_result = []
-        int8_postfix = "_eightbit"
-        for iter_res in activation_result:
-            tmp_iter_result = {}
-            for res in iter_res:
-                node_name, val = list(res.keys())[0], list(res.values())[0]
-                val = Dequantize(val[0], (node_name, val[1], val[2])) if len(val) == 3 else val
-                val = val.astype(np.float32)
-                index_postfix = node_name.find(int8_postfix)
-                if index_postfix != -1:
-                    node_name = node_name[:index_postfix]
-                    tmp_iter_result[node_name] = {node_name: val}
-                else:
-                    tmp_iter_result[fuse_map_reverse[node_name]] = {fuse_map_reverse[node_name]: val}
-            final_result.append(tmp_iter_result)
-        return final_result
-
-    def inspect_tensor(
-        self,
-        model,
-        dataloader=None,
-        op_list=[],
-        iteration_list=[],
-        inspect_type="activation",
-        save_to_disk=False,
-        save_path=None,
-        quantization_cfg=None,
-    ):
-        """Dump the weight and activation(output) to local disk.
-
-        1. create the correspondence between query node name and the actually output node name in graph_def
-        2. get the weight and bias for the given node
-        3. get the activation for the given node
-        4. save the tensor to disk
-        Args:
-            model: int8/fp32 graph_def/TensorflowBaseModel
-            dataloader: dataloader used during inspect activation
-            op_list: op list to inspect
-            iteration_list: iteration list to inspect, start from 1
-            inspect_type: activation/weight/all
-            save_to_disk: dump to disk or not
-            save_path: the dump path for inspect tensor
-            quantization_cfg: quantization configuration for fused fp32 model and quantized model
-        Returns:
-            Dict
-               {
-                 'weight': {
-                   'node0_name': {'weight0_name': numpy.array, 'bias0_name': numpy.array, ...},
-                   'node1_name': {'weight1_name': numpy.array, 'bias1_name': numpy.array, ...},
-                   ...
-                 },
-                 'activation': [
-                   # iter 1:
-                       {
-                         'node0_name': {'output0_name': numpy.array, 'output1_name': numpy.array, ...}
-                         'node1_name': {'output1_name': numpy.array, 'output1_name': numpy.array, ...}
-                         ...
-                       },
-                   # iter 2:
-                        {
-                       ...
-                       }
-                 ]
-               }
-        """
-        from neural_compressor.tensorflow.quantization.utils.graph_util import GraphAnalyzer
-        from neural_compressor.tensorflow.quantization.utils.utility import int8_node_name_reverse
-        from neural_compressor.tensorflow.utils import TensorflowBaseModel, dump_data_to_local, load_data_from_pkl
-
-        if isinstance(model, TensorflowBaseModel):
-            model = model.graph_def
-        if not quantization_cfg:
-            # TODO get config from graph if config is None
-            quantization_cfg = load_data_from_pkl("./nc_workspace/", "cfg.pkl")
-        node_list = op_list
-        # create the mapping between node name and node, key: node_name, val: node
-        graph_node_name_mapping = {}
-        quan_model_flag = False
-        for node in model.node:
-            node_name = int8_node_name_reverse(node)
-            if "Quantized" in node.op:
-                quan_model_flag = True
-                node_name = int8_node_name_reverse(node)
-            if node.attr["value"].tensor.dtype == tf.dtypes.bfloat16.as_datatype_enum:
-                quan_model_flag = True
-            graph_node_name_mapping[node_name] = node
-        if quan_model_flag:
-            logger.info("Dump the tensor for quantized model.")
-
-        # create the mapping between node name and node detail
-        g = GraphAnalyzer()
-        g.graph = model
-        graph_info = g.parse_graph()
-        inspect_result = {}
-
-        # inspect weight
-        if inspect_type == "weight" or inspect_type == "all":
-            logger.info("Start to inspect weight and bias.")
-            weights_result = self.inspect_weight_and_bias(node_list, model, graph_info, graph_node_name_mapping)
-            inspect_result["weight"] = weights_result
-
-        # inspect activation
-        if inspect_type == "activation" or inspect_type == "all":
-            logger.info("Start to inspect activation.")
-            activation_result = self.inspect_activation(
-                node_list, model, graph_node_name_mapping, quantization_cfg, dataloader, iteration_list, graph_info
-            )
-            inspect_result["activation"] = activation_result
-
-        # save to disk
-        if save_to_disk:
-            if not save_path:
-                save_path = "./nc_workspace/tmp/"
-            dump_data_to_local(inspect_result, save_path, "inspect_result.pkl")
-            logger.info(f"Dumped the inspect tensor to {save_path}")
-        return inspect_result
-
     def quantize_input(self, model):
         """Quantize the model to be able to take quantized input.
 
@@ -1381,320 +835,8 @@ class TensorFlowAdaptor:
                 res[op[1]] = {"activation": {"dtype": ["bf16"]}, "weight": {"dtype": ["bf16"]}}
         return res
 
-    def _pre_hook_for_qat(self, dataloader=None):
-        """Pre hook for QAT."""
-        self.model.model = self.qat_convert(self.model.model)
 
-    def _post_hook_for_qat(self):
-        """Post hook for QAT."""
-        pass
-
-    def _pre_eval_hook(self, model):
-        """Pre evaluation hook."""
-        return model
-
-    # Add keyword arguments unpacking
-    def _post_eval_hook(self, model, **kwargs):
-        """Post evaluation hook."""
-        pass
-
-    def save(self, model, path):
-        """Save model to the path."""
-        pass
-
-    # this function is used to convert keras QAT model to pb in old QAT implementation,
-    # and it's not used in refactored QAT
-    def convert(self, model, source, destination):  # pragma: no cover
-        """The function is used to convert a source model format to another.
-
-        Args:
-            model (neural_compressor.model): base model to be converted.
-            source (string): The source model format.
-            destination (string): The destination model format.
-        """
-        assert source.lower() == "qat" and destination.lower() == "default"
-        capability = self._query_fw_capability(model)
-
-        quantize_config = {"op_wise_config": {}}
-        for each_op_info in capability["opwise"]:
-            is_perchannel = False
-            weight_bit = 7.0
-            for op_cap in capability["opwise"][each_op_info]:
-                if "activation" in op_cap and "quant_mode" in op_cap["activation"]:
-                    activation = op_cap["activation"]
-                    if "weight" in op_cap:
-                        weight = op_cap["weight"]
-                        is_perchannel = True if weight["granularity"][0] == "per_channel" else False
-                    algorithm = activation["algorithm"][0]
-                    is_asymmetric = False
-                    if "activation" in op_cap:
-                        is_asymmetric = True if activation["scheme"][0] == "asym" else False
-
-                    quantize_config["op_wise_config"][each_op_info[0]] = (
-                        is_perchannel,
-                        algorithm,
-                        is_asymmetric,
-                        weight_bit,
-                    )
-        from neural_compressor.tensorflow.quantization.utils.graph_converter import GraphConverter
-
-        tmp_graphdef = copy.deepcopy(model.graph_def)
-        for i in tmp_graphdef.node:
-            if i.op == "Const" and i.input:
-                i.ClearField("input")
-        model.graph_def = tmp_graphdef
-        converter = GraphConverter(
-            model,
-            qt_config=quantize_config,
-            int8_sequences=self.op_wise_sequences,
-            fake_quant=True,
-            new_api=self.new_api,
-            performance_only=self.performance_only,
-            use_bf16=self.use_bf16,
-        )
-
-        return converter.convert()
-
-    def qat_convert(self, model, quantize_recipe=None):
-        """Convert a fp32 'tf.keras' model to be a int8 one with quantization aware training implementation.
-
-        Args:
-            model (tf.keras.Model): The model to be quantized, expected to be a Keras Functional or Sequential model.
-            quantize_recipe (dict): A dict that decide whether given layers should be quantized.
-
-        Returns:
-            converted_model (tf.keras.Model): Quantized model with fake quant nodes inserted.
-        """
-        assert isinstance(model, tf.keras.Model), (
-            "The model to be converted is expected to be "
-            "a `tf.keras.Model` instance. You should not pass an instance of type: {input}.".format(
-                input=model.__class__.__name__
-            )
-        )
-
-        assert model.__class__.__name__ in [
-            "Functional",
-            "Sequential",
-        ], "Only `Functional` or `Sequential` keras model is supported for QAT."
-
-        from neural_compressor.tensorflow.quantization.utils.quantize_graph.qat.quantize_config import global_config
-        from neural_compressor.tensorflow.quantization.utils.quantize_graph.qat.quantize_helper import (
-            init_quantize_config,
-            qat_clone_function,
-        )
-
-        config = init_quantize_config(model, quantize_recipe)
-        q_model = tf.keras.models.clone_model(model, input_tensors=None, clone_function=qat_clone_function)
-        global_config.clear()
-
-        return q_model
-
-    @dump_elapsed_time("Pass recover model")
-    def recover_tuned_model(self, model, q_config):
-        """Execute the recover process on the specified model.
-
-        Args:
-            tune_cfg (dict): quantization configuration
-            model (tf.compat.v1.GraphDef): fp32 model
-            q_config (dict): recover configuration
-
-        Returns:
-            tf.compat.v1.GraphDef: the quantized model
-        """
-        from neural_compressor.tensorflow.quantization.utils.graph_rewriter.generic.pre_optimize import PreOptimization
-
-        self.pre_optimizer_handle = PreOptimization(model, self.new_api, self.device)
-        self.pre_optimized_model = self.pre_optimizer_handle.get_optimized_model(self.itex_mode)
-        model.graph_def = self.pre_optimized_model.graph_def
-
-        from neural_compressor.tensorflow.quantization.utils.graph_converter_without_calib import (
-            GraphConverterWithoutCalib,
-        )
-
-        converter = GraphConverterWithoutCalib(
-            model,
-            recover_config=q_config,
-            new_api=self.new_api,
-            performance_only=self.performance_only,
-            use_bf16=self.use_bf16,
-        )
-
-        return converter.convert_without_calib()
-
-    def get_output_op_names(self, qmodel):
-        """Get the oupur OPs's names."""
-        from neural_compressor.tensorflow.quantization.utils.graph_util import GraphAnalyzer
-
-        graph_def = GraphAnalyzer().parse_graph(qmodel.graph_def)
-        output_op_names = set()
-
-        def _add_output_op_name(opname):
-            if opname.endswith("_dequantize"):
-                output_op_names.add(opname[: -len("_dequantize")])  # pylint: disable=no-member
-            elif opname.endswith("__dequant"):
-                pass
-            else:
-                output_op_names.add(opname)  # pylint: disable=no-member
-
-        for output_opname in qmodel.output_node_names:
-            op_count = 0
-            stack = [output_opname]
-            while stack:
-                opname = stack.pop()
-                while True:
-                    op_count += 1
-                    if opname not in graph_def:
-                        break
-                    op = graph_def[opname]
-                    if op.node.op == "Dequantize":
-                        _add_output_op_name(opname)
-                        break
-                    next_opnames = op.node.input
-                    if not next_opnames:
-                        break
-                    elif len(next_opnames) > 1:
-                        stack += next_opnames[1:]
-
-                    opname = next_opnames[0]
-
-        output_op_names = list(output_op_names)
-        logger.debug(f"output op names: {output_op_names}")
-        return output_op_names
-
-    def calculate_op_sensitivity(
-        self, model, dataloader, tune_cfg, output_op_names, confidence_batches, fallback=True, requantize_cfgs=None
-    ):
-        """Compute the op sensitivity.
-
-        The sensitivity metric is the mse between the output of the last quantized op of
-        the quantized model and the output of its corresponding op in the fp32 model.
-
-          1. Backup the tune cfg
-          2. Fallback each int8 op and compute its mse if use fallback (with 'fallback == True'),
-            or re-quantize each fp32 op(fallen back in the previous stage) and compute its MSE if not.
-          3. Sorted op name list according to its MSE
-
-        Args:
-          fp32_model: The fp32 model.
-          dataloader: the dataloader with full dataset.
-          tune_cfg: tuning config
-          fallback: denote fallback stage or re-quantize stage
-          requantize_cfgs: the dict of tuning configs for all re-quantizable ops
-
-        Returns:
-          A list of op names, sorted by its MSE sensitivity.
-        """
-        fp32_op_cfg = {"activation": {"dtype": "fp32", "quant_mode": "fp32"}, "weight": {"dtype": "fp32"}}
-
-        if fallback:
-            ops_list = [
-                op
-                for op, config in tune_cfg["op"].items()
-                if config["activation"]["quant_mode"] in ("static", "dynamic")
-            ]
-            replace_cfgs = {op: fp32_op_cfg for op in tune_cfg["op"]}
-        else:
-            ops_list = [
-                op
-                for op, config in tune_cfg["op"].items()
-                if config["activation"]["quant_mode"] == "fp32" and op in requantize_cfgs
-            ]
-            replace_cfgs = requantize_cfgs
-
-        # Step2. compute mse
-        mse_result = self._get_mse_order(
-            model, deepcopy(tune_cfg), replace_cfgs, ops_list, dataloader, output_op_names, confidence_batches
-        )
-
-        # Step3. sort
-        mse_order = [op for op, _ in sorted(mse_result.items(), key=lambda i: i[1])]
-        logger.debug("Dump MSE order:")
-        for op in mse_order:
-            logger.debug(f"{op}: {mse_result[op]}")
-        return mse_order
-
-    def _get_mse_order(
-        self, fp32_model, tune_cfg, replace_cfgs, ops_lst, dataloader, output_op_names, confidence_batches
-    ):
-        """Compute MSE."""
-        op_cfg = tune_cfg["op"]
-        mse_result = {}
-        partial_dataloader = self._partial_dataloader(dataloader, confidence_batches)
-
-        fp32_output = self._inference_model_on_batches(fp32_model, tune_cfg, partial_dataloader, output_op_names)
-
-        for op in ops_lst:
-            # backup and set replace tuning config
-            backup_cfg = op_cfg[op]
-            op_cfg[op] = replace_cfgs[op]
-
-            # quantize and inference the model
-            q_model = self.quantize(tune_cfg, fp32_model, partial_dataloader)
-            q_output = self._inference_model_on_batches(q_model, tune_cfg, partial_dataloader, output_op_names)
-
-            mse_result[op] = self._calculate_mse(fp32_output, q_output)
-
-            # recover tune_cfg
-            op_cfg[op] = backup_cfg
-
-        return mse_result
-
-    def _partial_dataset_of(self, dataloader, confidence_batches):
-        """Partial dataset."""
-        from neural_compressor.tensorflow.utils import DummyDataset, DummyDatasetV2
-
-        if isinstance(dataloader.dataset, DummyDataset) or isinstance(dataloader.dataset, DummyDatasetV2):
-            assert isinstance(confidence_batches, int)
-            ds = copy.deepcopy(dataloader.dataset)
-            ds.dataset = ds.dataset[:confidence_batches]
-            return ds
-        else:
-            return dataloader.dataset.take(confidence_batches)
-
-    def _partial_dataloader(self, dataloader, confidence_batches):
-        """Partial dataloader."""
-        return type(dataloader)(
-            dataset=self._partial_dataset_of(dataloader, confidence_batches),
-            batch_size=dataloader.batch_size,
-            last_batch=dataloader.last_batch,
-            collate_fn=dataloader.collate_fn,
-            sampler=dataloader.sampler,
-            batch_sampler=dataloader.batch_sampler,
-            num_workers=dataloader.num_workers,
-            pin_memory=dataloader.pin_memory,
-            shuffle=dataloader.shuffle,
-            distributed=dataloader.distributed,
-        )
-
-    def _calculate_mse(self, fp32_output, q_output):
-        """MSE calculation."""
-        result = []
-        for i, j in zip(fp32_output, q_output):
-            result.append(np.square(i - j).mean())
-        return np.array(result).mean()
-
-    def _inference_model_on_batches(self, model, tune_cfg, dataloader, output_op_names):
-        """Inference model on batches."""
-        from neural_compressor.tensorflow.quantization.utils.utility import generate_feed_dict
-
-        input_tensors = model.input_tensor
-        output_tensors = []
-        for op in output_op_names:
-            for tensor in model.graph.get_operation_by_name(op).outputs:
-                output_tensors.append(tensor)
-
-        predictions = []
-        for index, (inputs, _) in enumerate(dataloader):
-            feed_dict = generate_feed_dict(input_tensors, inputs)
-
-            pred = model.sess.run(output_tensors, feed_dict)
-            for item in pred:
-                predictions.append(item)
-
-        return predictions
-
-
-class Tensorflow_ITEXAdaptor(TensorFlowAdaptor):
+class Tensorflow_ITEXAdaptor(TensorFlowAdaptor):  # pragma: no cover
     """Tensorflow ITEX Adaptor Class."""
 
     def __init__(self, framework_specific_info):
@@ -2253,10 +1395,10 @@ class TensorflowQuery:
             if self.itex_mode:
                 patterns["int8"].append("FusedBatchNormV3 + Relu")
                 patterns["int8"].append("FusedBatchNormV3 + LeakyRelu")
-        elif version1_eq_version2(tf.version.VERSION, "1.15.0-up3"):
+        elif version1_eq_version2(tf.version.VERSION, "1.15.0-up3"):  # pragma: no cover
             patterns["int8"] = tf1_15_up3_int8_pattern_list
             patterns["uint8"] = tf1_15_up3_uint8_pattern_list
-        else:
+        else:  # pragma: no cover
             patterns["int8"] = old_tf_int8_pattern_list
             patterns["uint8"] = old_tf_uint8_pattern_list
