+++ neural_compressor/tensorflow/keras/layers/depthwise_conv2d.py
@@ -189,7 +189,7 @@ if version1_gte_version2(tf.__version__, "2.16.1"):
 
             return config
 
-else:
+else:  # pragma: no cover
 
     class QDepthwiseConv2D(DepthwiseConv):
         def __init__(
