+++ neural_compressor/tensorflow/keras/layers/separable_conv2d.py
@@ -198,7 +198,7 @@ if version1_gte_version2(tf.__version__, "2.16.1"):
 
             return config
 
-else:
+else:  # pragma: no cover
 
     class QSeparableConv2D(SeparableConv):
         def __init__(
