+++ neural_compressor/tensorflow/utils/__init__.py
@@ -46,14 +46,10 @@ from neural_compressor.tensorflow.utils.utility import (
     combine_histogram,
     get_all_fp32_data,
     get_tensor_histogram,
-    Dequantize,
-    dequantize_weight,
-    dump_data_to_local,
-    load_data_from_pkl,
     singleton,
     CpuInfo,
     Statistics,
     CaptureOutputToFile,
-    LazyImport,
     valid_keras_format,
+    TFSlimNetsFactory,
 )
