+++ neural_compressor/tensorflow/utils/utility.py
@@ -161,67 +161,6 @@ def get_tensor_histogram(tensor_data, bins=2048):
     return (hist, hist_edges, min_val, max_val, th)
 
 
-def Dequantize(data, scale_info):
-    """Dequantize the data with the scale_info."""
-    original_shape = data.shape
-    max_value = 255.0 if scale_info[0].find("Relu") != -1.0 else 127.0
-    _scale = (np.array(scale_info[2]) - np.array(scale_info[1])) / max_value
-    de_scale = np.ones(original_shape) * _scale
-    de_data = np.multiply(data, de_scale).astype(np.float32)
-    return de_data
-
-
-def dequantize_weight(weight_tensor, min_filter_tensor, max_filter_tensor):
-    """Dequantize the weight with min-max filter tensors."""
-    weight_channel = weight_tensor.shape[-1]
-    if len(min_filter_tensor) == 1:
-        weight_tensor = weight_tensor * ((max_filter_tensor[0] - min_filter_tensor[0]) / 127.0)
-    else:
-        # TODO to calculate the de-quantized result in a parallel way
-        for i in range(weight_channel):
-            weight_tensor[:, :, :, i] = weight_tensor[:, :, :, i] * (
-                (max_filter_tensor[i] - min_filter_tensor[i]) / 127.0
-            )
-    return weight_tensor
-
-
-def dump_data_to_local(data, path, filename):
-    """Dump data to local as pkl file.
-
-    Args:
-        data: Data used to dump
-        path: The directory to save data
-        filename: The filename to dump
-
-    Returns:
-        loaded data
-    """
-    from pathlib import Path
-
-    if not os.path.exists(path):
-        Path(path).mkdir(parents=True, exist_ok=True)
-    file_path = os.path.join(path, filename)
-    with open(file_path, "wb") as fp:
-        pickle.dump(data, fp)
-        logging.getLogger("neural_compressor").info("Dumped data to %s" % file_path)
-
-
-def load_data_from_pkl(path, filename):
-    """Load data from local pkl file.
-
-    Args:
-        path: The directory to load data
-        filename: The filename to load
-    """
-    try:
-        file_path = os.path.join(path, filename)
-        with open(file_path, "rb") as fp:
-            data = pickle.load(fp)
-            return data
-    except FileExistsError:
-        logging.getLogger("neural_compressor").info("Can not open %s." % path)
-
-
 def singleton(cls):
     """Not displayed in API Docs.
 
@@ -394,34 +333,143 @@ class CaptureOutputToFile(object):
         os.close(self.orig_stream_dup)
         self.tmp_file.close()
 
+@singleton
+class TFSlimNetsFactory(object):  # pragma: no cover
+    """TF-Slim nets factory."""
 
-class LazyImport(object):
-    """Lazy import python module till use."""
+    def __init__(self):
+        """Initialize a TFSlimNetsFactory."""
+        # tf_slim only support specific models by default
+        self.default_slim_models = [
+            "alexnet_v2",
+            "overfeat",
+            "vgg_a",
+            "vgg_16",
+            "vgg_19",
+            "inception_v1",
+            "inception_v2",
+            "inception_v3",
+            "resnet_v1_50",
+            "resnet_v1_101",
+            "resnet_v1_152",
+            "resnet_v1_200",
+            "resnet_v2_50",
+            "resnet_v2_101",
+            "resnet_v2_152",
+            "resnet_v2_200",
+        ]
 
-    def __init__(self, module_name):
-        """Init LazyImport object.
+        from tf_slim.nets import alexnet, inception, overfeat, resnet_v1, resnet_v2, vgg
+
+        self.networks_map = {
+            "alexnet_v2": {
+                "model": alexnet.alexnet_v2,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": alexnet.alexnet_v2_arg_scope,
+            },
+            "overfeat": {
+                "model": overfeat.overfeat,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": overfeat.overfeat_arg_scope,
+            },
+            "vgg_a": {
+                "model": vgg.vgg_a,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": vgg.vgg_arg_scope,
+            },
+            "vgg_16": {
+                "model": vgg.vgg_16,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": vgg.vgg_arg_scope,
+            },
+            "vgg_19": {
+                "model": vgg.vgg_19,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": vgg.vgg_arg_scope,
+            },
+            "inception_v1": {
+                "model": inception.inception_v1,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": inception.inception_v1_arg_scope,
+            },
+            "inception_v2": {
+                "model": inception.inception_v2,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": inception.inception_v2_arg_scope,
+            },
+            "inception_v3": {
+                "model": inception.inception_v3,
+                "input_shape": [None, 299, 299, 3],
+                "num_classes": 1001,
+                "arg_scope": inception.inception_v3_arg_scope,
+            },
+            "resnet_v1_50": {
+                "model": resnet_v1.resnet_v1_50,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": resnet_v1.resnet_arg_scope,
+            },
+            "resnet_v1_101": {
+                "model": resnet_v1.resnet_v1_101,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": resnet_v1.resnet_arg_scope,
+            },
+            "resnet_v1_152": {
+                "model": resnet_v1.resnet_v1_152,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": resnet_v1.resnet_arg_scope,
+            },
+            "resnet_v1_200": {
+                "model": resnet_v1.resnet_v1_200,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1000,
+                "arg_scope": resnet_v1.resnet_arg_scope,
+            },
+            "resnet_v2_50": {
+                "model": resnet_v2.resnet_v2_50,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": resnet_v2.resnet_arg_scope,
+            },
+            "resnet_v2_101": {
+                "model": resnet_v2.resnet_v2_101,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": resnet_v2.resnet_arg_scope,
+            },
+            "resnet_v2_152": {
+                "model": resnet_v2.resnet_v2_152,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": resnet_v2.resnet_arg_scope,
+            },
+            "resnet_v2_200": {
+                "model": resnet_v2.resnet_v2_200,
+                "input_shape": [None, 224, 224, 3],
+                "num_classes": 1001,
+                "arg_scope": resnet_v2.resnet_arg_scope,
+            },
+        }
+
+    def register(self, name, model_func, input_shape, arg_scope, **kwargs):
+        """Register a model to TFSlimNetsFactory.
 
         Args:
-           module_name (string): The name of module imported later
+            name (str): name of a model.
+            model_func (_type_): model that built from slim.
+            input_shape (_type_): input tensor shape.
+            arg_scope (_type_): slim arg scope that needed.
         """
-        self.module_name = module_name
-        self.module = None
-
-    def __getattr__(self, name):
-        """Get the attributes of the module by name."""
-        try:
-            self.module = importlib.import_module(self.module_name)
-            mod = getattr(self.module, name)
-        except:
-            spec = importlib.util.find_spec(str(self.module_name + "." + name))
-            mod = importlib.util.module_from_spec(spec)
-            spec.loader.exec_module(mod)
-        return mod
-
-    def __call__(self, *args, **kwargs):
-        """Call the function in that module."""
-        function_name = self.module_name.split(".")[-1]
-        module_name = self.module_name.split(f".{function_name}")[0]
-        self.module = importlib.import_module(module_name)
-        function = getattr(self.module, function_name)
-        return function(*args, **kwargs)
+        net_info = {"model": model_func, "input_shape": input_shape, "arg_scope": arg_scope}
+        net = {name: {**net_info, **kwargs}}
+        self.networks_map.update(net)
+        self.default_slim_models.append(name)
